requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.4.4
Flask-Login==0.5.0
Flask-WTF==0.15.1
Werkzeug==2.0.1
python-dotenv==0.19.0
email-validator==1.1.3
pycryptodome==3.19.1
SQLAlchemy==1.3.24
Markdown==3.3.4
Pygments==2.9.0
